pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "textual-haystack"
version = "0.1.0"
description = "Haystack integration for Tonic Textual PII detection, transformation, and entity extraction"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10.0,<4.0.0"
dependencies = [
    "haystack-ai>=2.0.0",
    "tonic-textual>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/haystack_integrations"]

[dependency-groups]
test = [
    "pytest>=7.3.0,<8.0.0",
    "pytest-asyncio>=0.21.1,<1.0.0",
]
lint = ["ruff>=0.13.1,<0.14.0"]
dev = ["haystack-ai"]
typing = ["mypy>=1.10.0,<2.0.0"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
    "integration: integration tests (require API key)",
]
